punto_1/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
pynamodb
python-dotenv